Added a slider for pitch bend. Created class ContinuiousFunction (maybe its useless). main.cpp is fucked up.

+++ util.cpp
@@ -7,7 +7,7 @@
 namespace util
 {
 
-Note::Note(float f)
+Note::Note(double f)
     :freq(f)
 {
